// ui/src/lib/api-client/models/GetMonitoringOauthEvents200ResponseEventsInner.ts
export interface GetMonitoringOauthEvents200ResponseEventsInner {
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    timestamp: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    type: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    clientId: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    clientName?: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    userName?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    scopes: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    grantType: string;
    /**
     * 
     * @type {number}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    responseTime: number;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    ipAddress: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    userAgent: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    errorMessage?: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    errorCode?: string;
    /**
     * 
     * @type {string}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    tokenType?: string;
    /**
     * 
     * @type {number}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    expiresIn?: number;
    /**
     * 
     * @type {boolean}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    refreshToken?: boolean;
    /**
     * 
     * @type {GetMonitoringOauthEvents200ResponseEventsInnerFhirContext}
     * @memberof GetMonitoringOauthEvents200ResponseEventsInner
     */
    fhirContext?: GetMonitoringOauthEvents200ResponseEventsInnerFhirContext;
}